% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakUpCluster.R
\name{BreakUpCluster}
\alias{BreakUpCluster}
\title{BreakUpCluster}
\usage{
BreakUpCluster(
  Object = Object,
  breakup.cluster = 1,
  sub.cluster = 3,
  uniquePathways = FALSE
)
}
\arguments{
\item{Object}{A PathwayObject}

\item{breakup.cluster}{The cluster to be broken up}

\item{sub.cluster}{The amount of subclusters to be generated}

\item{uniquePathways}{A boolean to specify the approach of unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.}
}
\value{
PathwayObject
}
\description{
Split a cluster in a desired number of subclusters.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)

GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")

OptimalGeneSets(Object = GSEA.Object3, uniquePathway = FALSE, cluster = 2,
                method = "silhouette", max_cluster= 24, cluster_method = "kmeans",
                main= "Kmeans for 24 clusters of Cluster 2 subclustering Gene-set")

GSEA.Object3.1 <- BreakUpCluster(GSEA.Object3, breakup.cluster=2, sub.cluster=2, uniquePathways=FALSE)
}

}
