% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterGeneSets.R
\name{ClusterGeneSets}
\alias{ClusterGeneSets}
\title{ClusterGeneSets}
\usage{
ClusterGeneSets(
  Object,
  clusters = 3,
  method = "kmeans",
  order = "group",
  molecular.signature = "All",
  user_function = NA
)
}
\arguments{
\item{Object}{A PathwayObject}

\item{clusters}{A numeric with the number of clusters required}

\item{method}{The cluster method specified for traditional clustering. Options are kmeans, kmeans_group,
Hierarchical and Hierarchical_group}

\item{order}{How should the data be ordered, by group or by cluster}

\item{molecular.signature}{to filter the molecular signatures which are duplicated
before clustering. options are Unique or All}

\item{user_function}{A user supplied function for clustering}
}
\value{
PathwayObject
}
\description{
Performs the traditional clustering in order to obtain groups of Gene-sets
and pathways based on distance similarity calculated prevously. Then,
performs the functional annotation from the generated clusters.
Traditional clustering supports: kmeans, hierarchical clustering and
function supplied by the user.
The clusters can be split up in subclusters using the mehtod BreakUpCluster.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)

GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                                clusters = 3,
                                method = "kmeans")
}

}
