% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterIndependentGeneSet.R
\name{ClusterIndependentGeneSet}
\alias{ClusterIndependentGeneSet}
\title{ClusterIndependentGeneSet}
\usage{
ClusterIndependentGeneSet(Object, use_method = NULL, nPathways = "optimal")
}
\arguments{
\item{Object}{PathwayObject.}

\item{use_method}{the method to use for ordering the correlation matrix.}

\item{nPathways}{a numeric or character. Optimal as default.}
}
\value{
PathwayObject
}
\description{
Performs seriation-based clustering of Gene-sets and pathways based on
previously calculated distance similarity. The function applies 29 different
seriation algorithms and selects the one with the highest score to represent
the ordering of the distance correlation matrix. Depending on the  minimum
number of Gene-sets estimated to form a cluster, the clusters will be defined.
The minimum number of Gene-sets is estimated automatically depending on the number
of Gene-sets and the number of clusters generated. But a custom minimum number
of Gene-sets can be selected using the method SetPathway.
}
\examples{

GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object2)
}

}
