% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineGeneSets.R
\name{CombineGeneSets}
\alias{CombineGeneSets}
\title{CombineGeneSets}
\usage{
CombineGeneSets(
  Object,
  combineMethod = "Standard",
  combineMethod.supplied,
  display = "Condensed",
  threads = 1
)
}
\arguments{
\item{Object}{A PathwayObject.}

\item{combineMethod}{lets the functions know if the standard RR, a Jaccard index or Cohens Kappa.}

\item{combineMethod.supplied}{a function which parameter A and parameter B.}

\item{display}{Either Condensed or Expanded. Expanded seperates the groups
and adds a 1 to every group is this molecular signature is seen here}

\item{threads}{Number of threads for parallelization.}
}
\value{
PathwayObject
}
\description{
Calculate distances between the different experiments. For both cases:
considering unique Gene-sets and unique Pathways.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\donttest{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
}
}
