% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGeneSets.R
\name{LoadGeneSets}
\alias{LoadGeneSets}
\title{LoadGeneSets}
\usage{
LoadGeneSets(
  file_location,
  groupnames,
  P.cutoff = 0.05,
  Mol.cutoff = 10,
  Source,
  Great.Background = F,
  type = NA,
  topranks = NA,
  structure,
  Organism,
  seperator
)
}
\arguments{
\item{file_location}{A list of GREAT, gseaResult or data.frame objects or location string in a vector of files.}

\item{groupnames}{A vector with group names of the different gene set experiments}

\item{P.cutoff}{numeric Pvalue cutoff for filtering.}

\item{Mol.cutoff}{numeric value for minimum number of molecules.}

\item{Source}{Tool used to generate gene sets.}

\item{Great.Background}{If the Great tool was used, did the user supply a background.}

\item{type}{For IPA data if "Canonical_Pathways" or "Functional_annotations" were supplied.}

\item{topranks}{numeric with the number of ranks per group to be loaded, usefull when there is a lot of data.}

\item{structure}{The structure of the genes. is it SYMBOLS, ENSEMBL, NCBI etc. Used for converting when there is mutiple structure in the object.}

\item{Organism}{the package name for the human or mouse data, used for converting the gene structure. name of the package, currently org.Hs.eg.db and org.Mm.eg.db supported.}

\item{seperator}{A character used within in the string to separate genes}
}
\value{
PathwayObject
}
\description{
Automatic loader for gene sets from the GSEA, GREAT or IPA web tools. The input can be either the file location
obtained from the online tools, or the GREAT or GSEA object obtained from the R tools. Also a template
data frame can be used to import Gene-Set analysis from other tools.
}
\examples{
# Loading from GREAT webtool results files:
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
                 system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"))


Great.Object1 <- LoadGeneSets(file_location = Great.files,
                              groupnames= c("KO", "WT"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 10,
                              Source = "Great",
                              Great.Background = FALSE,
                              type = "Canonical_Pathways",
                              topranks = 20,
                              structure = "SYMBOL",
                              Organism = "org.Mm.eg.db",
                              seperator = ",")

# Loading from IPA web tool results files:
IPA.files <- c(system.file("extdata",
                           "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls",
                            package = "GeneSetCluster"),
             system.file("extdata",
                            "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls",
                             package = "GeneSetCluster"),
             system.file("extdata",
                             "MM10.IPA.KO.uGvsMac.Functional_annotations.xls",
                             package = "GeneSetCluster"),
             system.file("extdata",
                             "MM10.IPA.WT.uGvsMac.Functional_annotations.xls",
                             package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files,
                         groupnames= c("KO", "WT"),
                         P.cutoff = 1.3,
                         Mol.cutoff = 5,
                         Source = "IPA",
                         type = "Canonical_Pathways",
                         structure = "SYMBOL",
                         seperator = ",",
                         Organism = "org.Mm.eg.db")

# Loading from GSEA results files:
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")


}
