% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalGeneSets.R
\name{OptimalGeneSets}
\alias{OptimalGeneSets}
\title{OptimalGeneSets}
\usage{
OptimalGeneSets(
  Object,
  uniquePathways = FALSE,
  cluster = NULL,
  method = "silhouette",
  max_cluster = 24,
  cluster_method = "kmeans",
  main = ""
)
}
\arguments{
\item{Object}{A PathwayObject.}

\item{uniquePathways}{A boolean to specify unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.
unique Pathway if it is TRUE.}

\item{cluster}{the cluster for sub-clustering.
If NA calculates the optimal for all the data.}

\item{method}{Which method to determing optimal number of clusters. gap, elbow or silhouette.}

\item{max_cluster}{Max number of clusters to test}

\item{cluster_method}{kmeans or hcut. Which clustering method is used}

\item{main}{A string to be used as title in the plot}
}
\value{
plot
}
\description{
There are many ways to determine what is the best separation of the clusters.
The function OptimalGeneSets can help with determining what is the best separation.
There are 3 different statistics the function can plot that help with this; Gap, Elbow and Silhouette.
Gap: Compares the total within intra-cluster variation for different values of k with their expected values under null reference distribution of the data
Elbow: For each k, calculate the total within cluster sum of squares
Silhouette: Determines how wel each object lies within its cluster. Higher the better.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)

OptimalGeneSets(Object = GSEA.Object2, uniquePathways = FALSE, cluster=NULL,
                method = "silhouette", max_cluster= 24, cluster_method = "kmeans",
                main= "Kmeans for 24 clusters")

OptimalGeneSets(Object = GSEA.Object2, uniquePathways = TRUE, cluster=NULL,
                method = "silhouette", max_cluster= 24, cluster_method = "kmeans",
                main= "Kmeans for 24 clusters")

GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")

OptimalGeneSets(Object = GSEA.Object3, uniquePathways = TRUE, cluster= 1,
                method = "silhouette", max_cluster= 24, cluster_method = "kmeans",
                main= "Kmeans for 24 clusters for Cluster 1")
}

}
