% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORAGeneSets.R
\name{ORAGeneSets}
\alias{ORAGeneSets}
\title{ORAGeneSets}
\usage{
ORAGeneSets(Object, uniquePathway = FALSE, clusterIndependent = FALSE)
}
\arguments{
\item{Object}{A PathwayObject.}

\item{uniquePathway}{A boolean to specify the approach of unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.}

\item{clusterIndependent}{A boolean to specify the clustering approach.
TRUE for seriation-based clustering approach and FALSE for classic approach.}
}
\value{
list with the over represented gene sets for all the clusters.
}
\description{
Obtains the results of the ORAs performed by the genes that make up each cluster.
}
\examples{

GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object=GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterIndependentGeneSet(Object=GSEA.Object2)
}
}
