% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayObject.R
\docType{class}
\name{PathwayObject-class}
\alias{PathwayObject-class}
\alias{PathwayObject}
\title{PathwayObject}
\value{
PathwayObject
}
\description{
Stores the all the information for GeneSetCluster. Essentially
pathways IDs, list of the genes, distance matrix and functional annotation
of the clusters.
}
\section{Slots}{

\describe{
\item{\code{Data}}{list. All the introduced data}

\item{\code{PData}}{data.frame.}

\item{\code{metadata}}{data.frame. Metdata of the object.}

\item{\code{plot}}{list. Information for plotting.}

\item{\code{Data.RR}}{data.frame. Distance matrix for unique Gene-sets.}

\item{\code{DataPathways.RR}}{data.frame. Distance matrix for unique Pathways.}

\item{\code{cIndependentMethod}}{list. Optimal seriation method for the data.}

\item{\code{dfTissue}}{list. Information about tissue enrichment in traditional clustering.}

\item{\code{dfTissueIndependent}}{list. Information about tissue enrichment in seriation-based clustering.}

\item{\code{functionalAnnot}}{list. Functional annotation about the groups generated in traditional clustering.}

\item{\code{functionalAnnotIndependent}}{list. Functional annotation about the groups generated in seriation-based clustering.}

\item{\code{stringdbInfo}}{list. Information of the STRING database and the parameters used.}
}}

