% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneNetworks.R
\name{PlotGeneNetworks}
\alias{PlotGeneNetworks}
\title{PlotGeneNetworks}
\usage{
PlotGeneNetworks(
  Object,
  uniquePathways = FALSE,
  clusterIndependent = FALSE,
  labels = FALSE,
  RRmin = 0.1
)
}
\arguments{
\item{Object}{A PathwayObject.}

\item{uniquePathways}{A boolean to specify the approach of unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.}

\item{clusterIndependent}{A boolean to specify the clustering approach.
TRUE for seriation-based clustering approach and FALSE for classic approach.}

\item{labels}{Boolean. TRUE to labels, or a vector with labels to be shown}

\item{RRmin}{filter for different RR cutoff for what is an edge. Minimum value 0 and Maximum value 1.}
}
\value{
plot
}
\description{
Plots a heatmap with the distances per cluster
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object3)

PlotGeneNetworks(GSEA.Object3, uniquePathways=FALSE, clusterIndependent=FALSE)
PlotGeneNetworks(GSEA.Object3, uniquePathways=FALSE, clusterIndependent=TRUE)
}
}
