% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneSets.R
\name{PlotGeneSets}
\alias{PlotGeneSets}
\title{PlotGeneSets}
\usage{
PlotGeneSets(Object, uniquePathways = FALSE, doORA = TRUE, wordcloud = TRUE)
}
\arguments{
\item{Object}{a PathwayObject}

\item{uniquePathways}{A boolean to specify the approach of unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.}

\item{doORA}{a boolean to show the ORA per cluster.}

\item{wordcloud}{a boolean to show the wordcloud annotation per cluster.}
}
\value{
plot
}
\description{
Plots the correlation matrix with the distances per cluster.
A functional annotation of the group of pathways based on GO terms descriptions.
Either Unique Gene-set or Pathway approach can be selected.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\donttest{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)

OptimalGeneSets(Object = GSEA.Object2, uniquePathways = FALSE, cluster=NULL,
                method = "silhouette", max_cluster= 24, cluster_method = "kmeans",
                main= "Kmeans for 24 clusters")

OptimalGeneSets(Object = GSEA.Object2, uniquePathways = TRUE, cluster = NULL,
                method = "silhouette", max_cluster= 24, cluster_method = "kmeans",
                main= "Kmeans for 24 clusters")

GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")

PlotGeneSets(GSEA.Object3, uniquePathways = FALSE)
PlotGeneSets(GSEA.Object3, uniquePathways = TRUE)
}

}
