% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathwayCluster.R
\name{PlotPathwayCluster}
\alias{PlotPathwayCluster}
\title{PlotPathwayCluster}
\usage{
PlotPathwayCluster(
  Object,
  doORA = TRUE,
  wordcloud = TRUE,
  uniquePathways = FALSE
)
}
\arguments{
\item{Object}{a PathwayObject}

\item{doORA}{a boolean to show the ORA per cluster.}

\item{wordcloud}{a boolean to show the wordcloud annotation per cluster.}

\item{uniquePathways}{A boolean to specify the approach of unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.}
}
\value{
plot
}
\description{
Plots the correlation matrix clustered by the seriation-based method.
A functional annotation of the group of pathways based on GO terms descriptions.
Either Unique Gene-set or Pathway approach can be selected.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

\donttest{
GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")

GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object2)
PlotPathwayCluster(GSEA.Object3, uniquePathways = FALSE)
PlotPathwayCluster(GSEA.Object3, uniquePathways = TRUE)
}

}
