% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSTRINGdbPerGeneSets.R
\name{plotSTRINGdbPerGeneSets}
\alias{plotSTRINGdbPerGeneSets}
\title{plotSTRINGdbPerGeneSets}
\usage{
plotSTRINGdbPerGeneSets(Object, plot.cluster = 1)
}
\arguments{
\item{Object}{a pathway object}

\item{plot.cluster}{The cluster for which the string image should be plotted.}
}
\description{
Plot the protein-protein interaction predcited using the STRINGdb plugin.
Before GetSTRINGdbPerGeneSets must be run.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\donttest{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object3)

GSEA.Object4 <- GetSTRINGdbPerGeneSets(Object = GSEA.Object3, uniquePathways=FALSE,
                                       clusterIndependent=FALSE)
GSEA.Object4 <- GetSTRINGdbPerGeneSets(Object = GSEA.Object4, uniquePathways=FALSE,
                                       clusterIndependent=TRUE)
plotSTRINGdbPerGeneSets(Object = GSEA.Object4, plot.cluster = 1)
plotSTRINGdbPerGeneSets(Object = GSEA.Object4, plot.cluster = 2)
plotSTRINGdbPerGeneSets(Object = GSEA.Object4, plot.cluster = 3)
}

}
