% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTissueExpression.R
\name{PlotTissueExpression}
\alias{PlotTissueExpression}
\title{PlotTissueExpression}
\usage{
PlotTissueExpression(
  Object,
  uniquePathways = FALSE,
  clusterIndependent = FALSE,
  showZscore = FALSE,
  all = FALSE
)
}
\arguments{
\item{Object}{a pathway object}

\item{uniquePathways}{A boolean to specify the approach of unique Gene-set or unique Pathways.
FALSE if for unique Gene-set and TRUE for unique Pathways.}

\item{clusterIndependent}{A boolean to specify the clustering approach.
TRUE for seriation-based clustering approach and FALSE for classic approach.}

\item{showZscore}{boolean to show Z score per cluster and in total}

\item{all}{boolean to show all tissues or only the top.}
}
\value{
plot
}
\description{
Plots a the tissue expression total median expression per tissue and the percentage of each cluster in each tissue
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object3)
GSEA.Object4 <- TissueExpressionPerGeneSet(GSEA.Object3, tissues="ALL", localDatabase=NULL, clusterIndependent=FALSE)
GSEA.Object4 <- TissueExpressionPerGeneSet(GSEA.Object4, tissues="ALL", localDatabase=NULL, clusterIndependent=TRUE)
PlotTissueExpression(GSEA.Object4, uniquePathways=FALSE, clusterIndependent=FALSE, showZscore=TRUE)
PlotTissueExpression(GSEA.Object4, uniquePathways=FALSE, clusterIndependent=TRUE, showZscore=TRUE)
}

}
