% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetPathway.R
\name{SetPathway}
\alias{SetPathway}
\title{SetPathway}
\usage{
SetPathway(Object, nPathways = "optimal")
}
\arguments{
\item{Object}{a PathwayObject.}

\item{nPathways}{a numeric or character. Optimal as default.}
}
\value{
PathwayObject
}
\description{
Set the minimum number of Gene-sets to be considered as group
in seriation-method clustering. Then the clusters will be definded and will
annotated.
}
\examples{

GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\donttest{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object2)
GSEA.Object3.1 <- SetPathway(GSEA.Object3, nPathways=12)
}

}
