% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayObject.R
\name{show,PathwayObject-method}
\alias{show,PathwayObject-method}
\title{show method for PathwayObject}
\usage{
\S4method{show}{PathwayObject}(object)
}
\arguments{
\item{object}{\code{\link{PathwayObject-class}} object}
}
\value{
summarize information about the object
}
\description{
Show method for a \code{\link{PathwayObject-class}} object
}
