% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueExpressionPerGeneSet.R
\name{TissueExpressionPerGeneSet}
\alias{TissueExpressionPerGeneSet}
\title{TissueExpressionPerGeneSet}
\usage{
TissueExpressionPerGeneSet(
  Object,
  tissues = "ALL",
  localDatabase = NULL,
  clusterIndependent = FALSE
)
}
\arguments{
\item{Object}{a PathwayObject.}

\item{tissues}{a character vector with the name of the tissues to use.}

\item{localDatabase}{the local GTEx database Object. If is NULL then the database will be obtained through custom API
Can be downloaded: https://github.com/TranslationalBioinformaticsUnit/GeneSetCluster}

\item{clusterIndependent}{A boolean to specify the clustering approach.
TRUE for seriation-based clustering approach and FALSE for classic approach.}
}
\value{
dataframe of pathwayobjects with tissue expression per cluster
}
\description{
Extracts every gen of every cluster from the object. Using the median gene
expression from the GTEx database build a data.frame with every tissue per cluster.
The list of available tissues:
Adipose...Subcutaneous; Adipose...Visceral..Omentum.; Adrenal.Gland; Artery...Aorta;
Artery...Coronary; Artery...Tibial; Bladder; Brain...Amygdala; Brain...Anterior.cingulate.cortex..BA24.;
Brain...Caudate..basal.ganglia.; Brain...Cerebellar.Hemisphere; Brain...Cerebellum; Brain...Cortex;
Brain...Frontal.Cortex..BA9.; Brain...Hippocampus; Brain...Hypothalamus; Brain...Nucleus.accumbens..basal.ganglia.;
Brain...Putamen..basal.ganglia.; Brain...Spinal.cord..cervical.c.1.; Brain...Substantia.nigra;
Breast...Mammary.Tissue; Cells...Cultured.fibroblasts; Cells...EBV.transformed.lymphocytes;
Cervix...Ectocervix; Cervix...Endocervix; Colon...Sigmoid; Colon...Transverse;
Esophagus...Gastroesophageal.Junction; Esophagus...Mucosa; Esophagus...Muscularis;
Fallopian.Tube; Heart...Atrial.Appendage; Heart...Left.Ventricle; Kidney...Cortex;
Kidney...Medulla; Liver; Lung; Minor.Salivary.Gland; Muscle...Skeletal; Nerve...Tibial;
Ovary; Pancreas; Pituitary; Prostate; Skin...Not.Sun.Exposed..Suprapubic.;
Skin...Sun.Exposed..Lower.leg.; Small.Intestine...Terminal.Ileum; Spleen; Stomach;
Testis; Thyroid; Uterus; Vagina; Whole.Blood.
}
\examples{
GSEA.files <- c(system.file("extdata", "Covid19AI_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid193Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"),
                system.file("extdata", "Covid196Mo_Healthy_GSEA.csv",
                             package = "GeneSetCluster"))

GSEA.Object1 <- LoadGeneSets(file_location = GSEA.files,
                              groupnames= c("IA", "Mo3", "Mo6"),
                              P.cutoff = 0.05,
                              Mol.cutoff = 5,
                              Source = "GSEA",
                              Great.Background = FALSE,
                              type = NA,
                              topranks = NA,
                              structure = "ENTREZID",
                              Organism = "org.Hs.eg.db",
                              seperator = "/")
\dontrun{
GSEA.Object2 <- CombineGeneSets(Object = GSEA.Object1, threads=2)
GSEA.Object3 <- ClusterGeneSets(Object = GSEA.Object2,
                clusters = 3,
                method = "kmeans")
GSEA.Object3 <- ClusterIndependentGeneSet(Object = GSEA.Object3)
GSEA.Object4 <- TissueExpressionPerGeneSet(GSEA.Object3, tissues="ALL", localDatabase=NULL, clusterIndependent=FALSE)
GSEA.Object4 <- TissueExpressionPerGeneSet(GSEA.Object4, tissues="ALL", localDatabase=NULL, clusterIndependent=TRUE)
}
}
