% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteGeneSets.R
\name{WriteGeneSets}
\alias{WriteGeneSets}
\title{WriteGeneSets}
\usage{
WriteGeneSets(
  Object,
  file_location = "~/Project9/test/",
  name = "IPA_20181123",
  write = "Both"
)
}
\arguments{
\item{Object}{is a PathwayObject.}

\item{file_location}{where to write the files to}

\item{name}{name to be added to the files, what experiments are these}

\item{write}{what to write, either "Data", "RR" or "Both"}
}
\value{
written tables in the folder designated.
}
\description{
Exports the pathways into a csv file
}
